spread out test only usage
