+++ tests/tt_metal/tt_metal/api/test_dram_to_l1_multicast.cpp
@@ -28,6 +28,7 @@
 #include <tt_stl/span.hpp>
 #include "tt_metal/test_utils/deprecated/tensor.hpp"
 #include <umd/device/types/arch.hpp>
+#include "impl/data_format/bfloat16_utils.hpp"
 
 using namespace tt;
